💄 style: fix formatting in gitlab.mk
The commit fixes the inconsistent date formats and indentation in the gitlab.mk file. The copyright year has been updated to include 2023. Additionally, the code now checks for two different tokens before using the credentials of the GitLab user that created the repository. Finally, the commit adds some debugging statements to print the values of `TF_USERNAME` and `TF_PASSWORD`.