✨ feat: Upgrade terraform packages
Upgraded the terraform packages by adding a new `upgrade` target in the Makefile. The upgrade target runs `terraform init -upgrade` with any optional arguments specified in `$(ARGS)`. This allows for easier management of terraform dependencies and staying up-to-date with the latest features and bug fixes.

Note: The `clean` target was also modified to include the `-x` flag for cleaning ignored files and directories.