✨ feat: Add aws_s3_bucket_public_access_block resource
This commit adds a new `aws_s3_bucket_public_access_block` resource to the `buckets.tf` file. This resource blocks public access to the bucket by setting `block_public_acls`, `block_public_policy`, and `restrict_public_buckets` parameters to `false`. Additionally, it includes a lifecycle block that allows for destruction of the resource. The `bucket` parameter is set to the ID of the `aws_s3_bucket.oidc` resource.

The commit was made on 26-06-2023 17:57:17 by Clay Risser.